% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_id.R
\name{api_screen_name}
\alias{api_screen_name}
\title{Get the screen name of the authenticated user}
\usage{
api_screen_name(token = NULL)
}
\arguments{
\item{token}{Token}
}
\description{
Get the screen name of the authenticated user
}
