% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geocode_request}
\alias{geocode_request}
\title{Find geo-locations searched from a string}
\usage{
geocode_request(string = "Lagos, Nigeria",
  token = Sys.getenv("LOCATIONIQ_TOKEN"))
}
\arguments{
\item{string}{String. The string to search in the API. Default = "Lagos, Nigeria"}

\item{token}{String. The token to use to make the API call. Default = Sys.getenv('TWITTER_CONSUMER_API_KEY')}
}
\value{
Dataframe of the response from the API.
}
\description{
This uses the free api from locationIQ. It is free to get a token and make a call to this API without the need for approval.
(Note the rate limit for this API is 2 seconds)
}
\details{
Most twitter users haven't opted in to using location data for their tweets, but many have location descriptions.
This function is a helper to scrape those locations.
}
\examples{
response <- geocode_request('Big Ben, London')

}
\seealso{
Visit \href{https://locationiq.com/register}{LocationIQ website} to sign up for a free token.
}
\keyword{api}
\keyword{call}
\keyword{generic}
\keyword{twitter}
