% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shinyTwitterApp}
\alias{shinyTwitterApp}
\title{Twitter scraping shiny app}
\usage{
shinyTwitterApp(port = 5000, host = "127.0.0.1")
}
\arguments{
\item{port}{Integer: Port to run the shiny app through. Defaults to 5000.}

\item{host}{String: IP address to host the shiny app. Defaults to 0.0.0.0.}
}
\description{
This function runs an example shiny app for twitter scraping.
}
\keyword{shiny}
\keyword{twitter}
