% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{TWIT}
\alias{TWIT}
\alias{TWIT_get}
\alias{TWIT_post}
\title{Send POST and GET requests to the Twitter API}
\usage{
TWIT_get(token, api, params = NULL, ..., host = "api.twitter.com")

TWIT_post(
  token,
  api,
  params = NULL,
  body = NULL,
  ...,
  host = "api.twitter.com"
)
}
\arguments{
\item{token}{the token}

\item{api}{which API are we using}

\item{params}{parameters for the query}

\item{...}{parameters to be passed to httr::GET or httr::POST}

\item{host}{the API host}
}
\description{
Send POST and GET requests to the Twitter API
}
