% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_id.R
\name{user_type}
\alias{user_type}
\title{Checks if user information is a user_id or screen_name}
\usage{
user_type(x, arg_name = "user")
}
\arguments{
\item{x}{A twitter user_id or screen_name}

\item{arg_name}{Name of the argument to return in error message}
}
\value{
Either "user_id" or "screen_name"
}
\description{
Checks if user information is a user_id or screen_name
}
