% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{twitter_oauth2}
\alias{twitter_oauth2}
\title{Authenticate Twitter session using OAuth 2.0}
\usage{

  twitter_oauth2(consumer_api_key = Sys.getenv("TWITTER_CONSUMER_API_KEY"),
  consumer_secret_api_key = Sys.getenv("TWITTER_CONSUMER_SECRET_API_KEY"))
}
\arguments{
\item{consumer_api_key}{String. The consumer api key provided for using your app. Default = Sys.getenv('TWITTER_CONSUMER_API_KEY')}

\item{consumer_secret_api_key}{String. The consumer api secret key provided for using your app. Default = Sys.getenv('TWITTER_CONSUMER_SECRET_API_KEY')}
}
\description{
This function creates a hidden environment that holds the authentication token required for making calls to the Twitter API.
}
\details{
To use the twitter api, you will need to an approved account with an app (that has both an api key and a secret api key)
You can find your approved apps at \url{https://developer.twitter.com/en/apps/} (if logged in)
This function is called when running other functions (such as \code{\link{generic_api_call}})
}
\examples{
twitter_oauth2() # by default,
# twitter_oauth2('<YOUR API KEY>', '<YOUR SECRET API KEY>') # Not run - input your own api keys

}
\seealso{
\url{https://stackoverflow.com/a/32534239}
}
\keyword{api}
\keyword{authenticate}
\keyword{twitter}
